Handle HTTP POST requests with raw form data
A recent version of openwpm-webext-instrumentation also supports HTTP
POST request in which the POST body is only available as raw binary
data that is stored in the post_body_raw column.

This change updates the database structure with that column and adjusts
the tests so that those cases are correctly tested.